% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh-package.R
\docType{package}
\name{feh-package}
\alias{feh}
\alias{feh-package}
\title{feh: Fire Emblem Heroes Summoning Simulator}
\description{
Produces orb cost distribution based on a user's desired outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nelschua.github.io/feh-summoning/}
}

}
\author{
\strong{Maintainer}: Nelson Chua \email{nelson.chua@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
