% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\title{Generate summoning session}
\usage{
session(herocounts, baserate, rateups, targetcol, targetrar)
}
\arguments{
\item{herocounts}{4x4 matrix of containing colour x rarity counts}

\item{baserate}{Decimal for the initial 5* appearance rate}

\item{rateups}{Integer for the number of times pity has increased}

\item{targetcol}{Colour of desired character}

\item{targetrar}{Rarity of desired character}
}
\value{
5x3 matrix with columns Rarity, Colour and Target
}
\description{
Generate summoning session
}
\examples{
session(matrix(c(15,27,25,0,12,23,17,1,9,18,12,1,13,23,16,2),4,4,byrow=TRUE), 0.03, 0, "Red", "5*F")
}
